third_party/WebKit/LayoutTests/http/tests/security/cross-origin-OffscreenCanvas2D-transferToImageBitmap-expected.txt

The taintedness of the imagebitmap drawn to OffscrenCanvas must be transfered

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ImageBitmap is tainted. Threw error: SecurityError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The canvas has been tainted by cross-origin data.
PASS successfullyParsed is true

TEST COMPLETE
